# Services/customer-service/python-flask-server/swagger_server/models/card_list.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.card import Card  # noqa: F401,E501
from swagger_server import util


class CardList(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self):  # noqa: E501
        """CardList - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'CardList':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CardList of this CardList.  # noqa: E501
        :rtype: CardList
        """
        return util.deserialize_model(dikt, cls)